Add donation handler endpoints
Move a bunch of files around, with the hopeful end result of making it
easier to understand. Instead of using the Stripe API library, I just
wrapped a bunch of API calls in helper functions that return promises.

We want people to be able to give any amount (USD) as a one time
donation or on a monthly basis. So the donate endpoint should handle
both of those cases instead of the previous way of using two different
endpoints.

The endpoint needs an amount (USD), an email, an indication of if the
donation is recurring and a token from Stripe. It should pass along the
successful API response or some sort of error for the client to handle.